Solving issues in the domain models

+++ commercetools.Sdk/commercetools.Sdk.Domain/Query/TermFacetResult.cs
@@ -6,9 +6,9 @@ namespace commercetools.Sdk.Domain
     public class TermFacetResult : FacetResult
     {
         public string DateType { get; set; }
-        public int Missing { get; set; }
-        public int Total { get; set; }
-        public int Other { get; set; }
+        public long Missing { get; set; }
+        public long Total { get; set; }
+        public long Other { get; set; }
         public List<FacetTerm> Terms { get; set; }
     }
 }
