+++ commercetools.Sdk/commercetools.Sdk.Domain/Reviews/ReviewDraft.cs
@@ -16,7 +16,7 @@ namespace commercetools.Sdk.Domain.Reviews
         public string Title { get; set; }
         public string Text { get; set; }
         public IReference Target { get; set; }
-        public double? Rating { get; set; }
+        public int? Rating { get; set; }
         public IReference<State> State { get; set; }
         public IReference<Customer> Customer { get; set; }
         public CustomFields Custom { get; set; }
